tfjs/requirements.txt
onnx==1.10.1
onnx-tf==1.9.0
segmentation-models-pytorch==0.2.0
tensorflowjs==3.9.0
